refactor(web): organize user payment methods
